requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3